ability to query for dkg contract state
