+++ contracts/coconut-dkg/src/dealings/transactions.rs
@@ -6,7 +6,7 @@ use crate::dealings::storage::StoredDealing;
 use crate::epoch_state::storage::{CURRENT_EPOCH, INITIAL_REPLACEMENT_DATA};
 use crate::epoch_state::utils::check_epoch_state;
 use crate::error::ContractError;
-use crate::state::STATE;
+use crate::state::storage::STATE;
 use cosmwasm_std::{DepsMut, MessageInfo, Response};
 use nym_coconut_dkg_common::types::{EpochState, PartialContractDealing};
